Fix fetchall_arrayref({}) behavior with no columns
When no columns were present in a result set, fetchall_arrayref({}) died
because the resulting arguments to bind_columns were incorrect (by
virtue of not existing at all). Since there can't be results with an
empty column set, this fixes the issue by returning an empty result set
in that case.

This issue was first referenced in November 2012:
http://www.nntp.perl.org/group/perl.dbi.users/2012/11/msg36513.html